fix: exclusion weight where type is not in the excluded relation (#443)
For relation A but not B, we need to use a mixed strategy in calculating
weight where a type in A will use the max strategy and B will use the
enforced type strategy. This is because a type will be connected to the
graph if it appears in A as B is not mandatory.  However, a type listed
in B MUST also appear in A.  Otherwise, relation will never exist for such type.